feat(frontend): add CanteenSelection component
